Remove default title in issue templates
The default title prefixes the issue title with `[BUG]` or `[FEAT]`. If there is a reason to do this, feel free to just close the PR, however I see no reason that this should be here considering we have labels for filtering. The `bug` and `enhancement` labels are automatically applied when using the issue templates.